// wdip-fe/src/service/wdip-be/models/MotionsByPartyResults.ts
// tslint:disable
/**
 * WDIP
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * OpenAPI spec version: 0.0.1
 * Contact: markus@silberstein.nu
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists } from '../runtime';
/**
 * 
 * @export
 * @interface MotionsByPartyResults
 */
export interface MotionsByPartyResults {
    /**
     * 
     * @type {string}
     * @memberof MotionsByPartyResults
     */
    party?: string;
    /**
     * 
     * @type {number}
     * @memberof MotionsByPartyResults
     */
    submitted?: number;
    /**
     * 
     * @type {number}
     * @memberof MotionsByPartyResults
     */
    approved?: number;
    /**
     * 
     * @type {number}
     * @memberof MotionsByPartyResults
     */
    declined?: number;
}

export function MotionsByPartyResultsFromJSON(json: any): MotionsByPartyResults {
    return {
        'party': !exists(json, 'party') ? undefined : json['party'],
        'submitted': !exists(json, 'submitted') ? undefined : json['submitted'],
        'approved': !exists(json, 'approved') ? undefined : json['approved'],
        'declined': !exists(json, 'declined') ? undefined : json['declined'],
    };
}

export function MotionsByPartyResultsToJSON(value?: MotionsByPartyResults): any {
    if (value === undefined) {
        return undefined;
    }
    return {
        'party': value.party,
        'submitted': value.submitted,
        'approved': value.approved,
        'declined': value.declined,
    };
}